Major code refactoring, developing mirroring functionality
