+++ src/__init__.py
@@ -1,5 +1,10 @@
 ### General libraries ###
-# import pip # NOT NEEDED FOR THIS VERSION
+from .main_gui.panel import MainPanel
+from .main_gui.decompose_button import DecomposeButton
+from .main_gui.export_button import ExportButton
+from .mirroring.flip_button import FlipButton
+from .mirroring.mirror_menu import MirrorMenu
+from .strings.keymaps import keymaps
 
 ### Blender libraries ###
 import bpy
@@ -10,7 +15,13 @@ _props = []
 Add-on properties
 '''
 
-_classes = []
+
+_classes = [MainPanel, DecomposeButton, ExportButton, FlipButton, MirrorMenu]
+'''
+Add-on classes
+'''
+
+_keymaps_list = []
 
 # _modules = ["debugpy"]
 # '''
@@ -21,10 +32,10 @@ _classes = []
 bl_info = {
  "name": "MyoGenerator Exporter",
  "description": "Addon used for exporting muscle models created by the Myogenerator, decomposing them into muscle fibers using 3rd party software and then importing them back to Blender",
- "author": "Jan Heres",
- "blender": (2, 91, 0),
- "version": (0, 0, 1),
- "warning": "Only officially supported for versions 2.93.X of Blender!",
+ "author": "Jan Hereš (www.janheres.eu)",
+ "blender": (2, 93, 0),
+ "version": (1, 0, 1),
+ "warning": "Only officially supported for versions 2.91.X - 2.93.X of Blender!",
  "category": "Export",
 }
 '''
@@ -32,20 +43,18 @@ Add-on info in Edit -> Preferences -> Add-ons
 '''
 
 def register():
+    register_classes()
+    register_props()
+    register_keymaps()
+
+def register_classes():
     global _classes
-
-    # install_modules() # Making sure the essential modules are installed - NOT NEEDED IN THIS VERSION
-
-    # NOTE: Have to import the add-on's classes AFTER the essential modules are installed
-    from .addon_panel import MainPanel
-    from .addon_decompose_button import DecomposeButton
-    from .addon_export_button import ExportButton
-    
-    _classes = [MainPanel, DecomposeButton, ExportButton]
 
     for cls in _classes:
         register_class(cls)
 
+
+def register_props():
     bpy.types.Scene.output_path = bpy.props.StringProperty(
         name = "Output path (folder)",
         default = "",
@@ -96,6 +105,59 @@ def register():
     _props.append(bpy.types.Scene.export_resolution)
     _props.append(bpy.types.Scene.export_visualize)
 
+def register_keymaps():
+    global _keymaps_list
+
+    wm = bpy.context.window_manager
+    kc = wm.keyconfigs.addon
+
+    for keymap_group_name in keymaps:
+        keylist = keymaps[keymap_group_name]
+
+        for list_item in keylist:
+            menu_name = list_item.get("menu_name")
+            space_type = list_item.get("space_type", "EMPTY")
+            region_type = list_item.get("region_type", "WINDOW")
+
+            if list_item:
+                km = kc.keymaps.new(menu_name, space_type=space_type, region_type=region_type)
+
+                if km:
+                    action = list_item.get("action")
+                    idname = list_item.get("idname")
+                    type = list_item.get("type")
+                    value = list_item.get("value")
+
+                    shift = list_item.get("shift", False)
+                    ctrl = list_item.get("ctrl", False)
+                    alt = list_item.get("alt", False)
+                    oskey = list_item.get("oskey", False)
+
+                    kmi = km.keymap_items.new(action, type, value, shift=shift, ctrl=ctrl, alt=alt, oskey=oskey)
+
+                    if kmi:
+                        kmi.properties.name = idname
+                _keymaps_list.append(km)
+
+    # Mirroring keymap group 
+    # km = wm.keyconfigs.addon.keymaps.new("Window", space_type="EMPTY", region_type="WINDOW")
+
+    # # Particular key shortcuts
+    # kmi = km.keymap_items.new("wm.call_menu", type="F5", value="PRESS")
+    # kmi.properties.name = MirrorMenu.bl_idname
+    # _keymaps_list.append(km)
+
+
+def unregister_keymaps():
+    global _keymaps_list
+
+    wm = bpy.context.window_manager
+
+    for km in _keymaps_list:
+        for kmi in km.keymap_items:
+            km.keymap_items.remove(kmi)
+    _keymaps_list.clear()
+
 
 def unregister():
     for cls in _classes:
@@ -104,8 +166,10 @@ def unregister():
     for prop in _props:
         del prop
 
+    unregister_keymaps()
 
-# NOT NEEDED FOR THIS VERSION
+
+# DEPRECATED
 # '''
 # Installing all dependent modules
 # '''
