+++ src/strings/descriptions.py
@@ -1,0 +1,8 @@
+'''
+Strings as bl_descriptions of each element
+'''
+descriptions = {
+    "ExportButton_bl_description" : "Export selected muscle parts (volume, origin, insertion)",
+    "DecomposeButton_bl_description" : "Decompose specified/all muscle(s) in the selected directory",
+    "FlipButton_bl_description" : "Flips selected object(s) by a selected axis",
+}
