+++ src/strings/errors.py
@@ -1,0 +1,15 @@
+errors = {
+    "Message_nothing_selected" : "No object(s) selected!",
+    "Message_exporting_finished" : "Exporting finished!",
+    "Message_wrong_name_format" : "Wrong muscle name format! Model should be named using following pattern: <muscle_name><space><origin|volume|insertion><space><boundary>!",
+    "Message_decomposing_done" : "Decomposing done!",
+    "Message_muscle_not_found" :"Corresponding muscle not found. Please, adjust the model manually",
+    "Message_file_not_found" : "File was not found! Please, check the filename and directory.",
+    "Message_nothing_to_export" : "No muscles to export found in the specified directory.",
+    "Message_mdt_not_found" : "MuscleDecomposition executable has not been found in the add-on folder! Please, proceed to user manual for further information.",
+    "Message_decomposition_unsuccessful" : "Decomposition has been aborted due to the following reason: ",
+    "Message_decomposition_done" : "MuscleDecompositionTest output: ",
+}
+'''
+Strings for possible errors that might occur in this add-on
+'''
