+++ src/strings/keymaps.py
@@ -1,0 +1,20 @@
+from ..mirroring.mirror_menu import MirrorMenu
+
+keymaps = {"Mirroring": [{
+                "label": "Mirror object(s)",
+                "region_type": "WINDOW",
+                "space_type": "EMPTY",
+                "map_type": "KEYBOARD",
+                "menu_name": "Window",
+                "action" : "wm.call_menu",
+                "idname": MirrorMenu.bl_idname,
+                "type": "F5",
+                "ctrl": False,
+                "alt": False,
+                "shift": False,
+                "oskey": False,
+                "value": "PRESS"
+            }]}
+'''
+Add-on specific keymaps
+'''
