+++ src/strings/labels.py
@@ -1,0 +1,15 @@
+'''
+Strings as bl_labels of each element
+'''
+labels = {
+    "MainPanel_bl_label" : "MyoGenerator Exporter",
+    "MirrorMenu_bl_label" : "Mirror menu",
+    "ExportButton_bl_label" : "Export selected muscle(s)",
+    "DecomposeButton_bl_label_specified" : "Decompose specified muscle",
+    "DecomposeButton_bl_label_all" : "Decompose all muscles",
+    "FlipButton_bl_label" : "Flip selected object(s)",
+    "ProgressPopup_bl_label" : "Exporting successful!",
+    "Label_output_path_text" : "Output directory",
+    "Label_output_filename_text" : "Muscle name (decomposition)",
+    "Label_parameters_text" : "Muscle decomposition parameters",
+}
