+++ src/strings/props.py
@@ -1,0 +1,7 @@
+'''
+Strings for different props
+'''
+props = {
+    "DecomposeButton_type_specified" : "specified",
+    "DecomposeButton_type_all" : "all",
+}
